feat(kernel): Adds SBI mod
SBI (Supervisor Binary Interface) is an abstraction layer between the machine and kernel that offers a set of higher-level APIs we can interface with so that the kernel doesn't have to.

One such API is HSM (Hart State Management) that lets us start, suspend and shutdown individual harts.
As per SBI-spec all harts except for the boot hart are put into suspension when entering the kernel and need to be woken through calling the HSM `start_hart` function.