feat(kernel): trap handling
Traps (Interrupts and Exception) require us to handle them by specifying an address to jump to when a trap is taken.

We use the *vectored* trap mode which will allow us to specify fast path handlers for certain traps.
Falling back to the general case handler which will save all `a` and `t` registers plus the `ra` register into a `TrapFrame` struct, and restores them after the trap handler completes.

To work around having to allocate the trap frame anywhere we simply push it onto the stack as the very first stack item.

Currently the harts stack also gets reused for the trap handler, this helps us not having to do expensive stack switching.