feat(kernel): implement sv39 virtual memory scheme
This implements the most common RiscV virtual memory scheme, sv39.

Implementation includes a mapper, with a `map` method to map a virtual address to a physical one and a `virt_to_phys` method to manually do the address translation.
None of the code is tested yet and error handling as well as performance can definetely be improved, *but* identity-mapping the kernel & switching on the MMU works at least so this is a good starting point.

This also temporarily removes the `kmem` heap allocation code because that needs adapting to the new virtual memory system