fix(kernel) disable interrupts while in logging
When the trap handler gets invoked the kernel code that was running at the time essentially gets frozen. This means that if the kernel code was holding a lock at the time of being frozen, the trap handler would deadlock as it would attempt wait forever to aquire that lock.

We fix this by temporarily disabling interrupts in while logging.