refactor: `Mapper` uses borrowed `FrameAllocator`
This change makes the `Mapper` use a mutable reference to the `FrameAllocator` instead of owning it. It also stores the `FrameAllocator` as a global object instead of the mapper and turning the mapper into a short-lived object.

This is done so we can reuse the frame allocator for userspace mapping too