fix(thread_local): work around compiler bug by (falsely) marking constant static as mutable
The `thread_local` macro has an optimization case for constant initilalized locals, where we skip a lot of the ceremony. But due to (likely) a bug in the Rust compiler const thread local statics data would be placed into tdata even if it is all zeros.

Marking the static as `mut` forces the compiler to skip the shortcut and correctly place the data into tbss