fix(kernel): fix dynamic allocator mapping
The kernel allocator dynamically maps in more backing memory as needed, however previously this was broken as we tried to access the memory before mapping it.
This change addresses this as well as changing it so that the heap is mapped in from the bottom addresses upwards instead of the other way around.