fix(kernel): use `Vec::extend_from_slice` instead of raw `copy_nonoverlaping`
Memory initialization would fail since `Vec` doesn't allocate when constructed with a capacity of zero. This would result in a dangling pointer when accessed through `Vec::as_mut_ptr` and that being written to which caused  a trap.
The correct behaviour (and the one introduced by this fix) is to use the high level `Vec::extend_from_slice` since that handles allocation under the hood.