refactor: simplify higher-half mapping logic
The previous commit change the virtual address allocation scheme so that the loader would only allocate kernel regions in the higher half of memory (the "kernel space").

That implementation was quite simple naive though, since it just marked half the address space as used.

This change modifies implementation so that it only keeps track of the upper half, which cuts stack consumption in half and also makes the calculation of allocated addresses easier since we don't need to sign extend anymore.