chore: vendor linked list allocator (#101)
* fix(vmm): correctly map additional BSS pages (#73)

This change corrects a bug in the `vmm` elf loading logic, where BSS sections larger than one page would not be correctly mapped.

* chore(deps): update rust crate syn to v2.0.74 (#75)

* fix(deps): update rust crate serde to v1.0.206 (#76)

* fix(deps): update rust crate serde to v1.0.207 (#77)

* chore(deps): update rust crate clap to v4.5.15 (#74)

* refactor: remove vendored gimli copy (#78)

* refactor: remove vendored gimli copy

* cleanup

* fix(kstd): use `payload_as_str` function to correctly print panic message in recursive panics

* feat: pass `payload_elf_phys` info to payload

This change passes information about the physical memory region occupied by the payload to the payload itself. This will be used by the kernel to parse its own elf file and find debugging info

* cleanup

* clippy

* fix(deps): update rust crate serde to v1.0.208 (#81)

* chore(deps): update rust crate clap to v4.5.16 (#82)

* chore(deps): update rust crate arrayvec to v0.7.6 (#83)

* chore(deps): update rust crate syn to v2.0.75 (#84)

* chore(deps): update rust crate postcard to v1.0.9 (#86)

* chore: build setup & tooling (#85)

This change replaces the old build setup which has become a bit of a liability with a more streamlined, easier to maintain and incidentally even faster one using just, nushell and optionally nix for a dev environment.

* add boot flow docs (#80)

* Update README.md (#52)

* Update README.md

* Update README.md

* Update README.md

* Update roadmap.md

* Update SUMMARY.md

* fix: allow QEMU to be interrupted with ctrl+c

* chore(deps): update actions/checkout action to v4 (#88)

* chore(deps): update rust crate postcard to v1.0.9 (#87)

* Update justfile

* refactor: break up kstd crate into multiple smaller crates (#90)

This change breaks up the monolithic `kstd` crate into multiple smaller crates (`panic`, `unwind`, `sync`, `backtrace`, `tls`, and `riscv`) that are easier to compose, work on and extend

* Update SUMMARY.md

* feat(loader): Support Position Independent Executables (#93)

* chore(deps): update rust crate syn to v2.0.76 (#92)

* disable verbose relocating logs

* feat: Kernel Address Space Layout Randomization (KASLR) (#94)

This change builds off of the support for position independent code added in #93 to implement KASLR in the loader crate.

This is done by using the rng-seed provided by the FDT (i.e. by the previous stage loader) to seed a ChaCha20-based random number generator. The generator is then used to generate a cryptographically secure base address for the payload.
Note that a page-aligned address is chosen in a window between PHYS_OFFSET and the highest possible, i.e. PHYS_OFFSET..(u64::MAX - kernel_size) which is around 2^25 depending on the kernel size (the entropy will get smaller as the kernel grows).

* Update README.md

* Update README.md (#95)

* cleanup

* refactor: remove backtrace feature from panic (#96)

This change removes the backtrace feature from the `panic` crate and the backtrace printing from the default panic hook. Instead a custom panic hook should be used.

* refactor: remove hardcoded `PHYS_OFFSET` constant (#98)

This change removes the hardcoded `PHYS_OFFSET` constant from the `kmm` crate making it possible to configure the location of the physical memory mapping at runtime.

Actually making the position dynamic (randomized as part of KASLR is left for a later change)

* feat(loader-api): pass physical payload location (#97)

This changes passes information about the location of the payload in physical memory to the payload itself so it can perform introspection into its own elf file if desired (e.g. to access eh_frame info for stack unwinding for debug info for symbolization)

* fix: add physmem off to `kaslr:: random_offset_for_payload`

* feat: improve KASLR (#99)

* Update README.md

* Update intro.md

* refactor(loader): rename all occurrences of `payload` to `kernel` for clarity

* fix typo

* fix rust targets

* cleanup and docs

* chore: vendor `linked-list-allocator`

The `linked-list-allocator` crate is pretty essential to the inner workings of k23, being the allocator of choice when a simple inexpensive heap is required. This change brings the `linked-list-allocator` crate in-tree to better tailor it to our needs and eliminate 3rd party dependencies

* clippy & fixes

* Update Cargo.toml

* typos

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>