feat(kernel/vm): implement efficient gap search state
This change uses the `wavltree` lifecycle hooks to maintain an augmented version of the WAVLTree that supports more efficient gap search. This is inspired by a similar inspiration in the Zircon kernel and works by tracking the largest and smallest address of each nodes subtree, as well as the biggest gap in each subtree.