feat(kernel/vm): allocate virtual memory spot
Implement the `find_spot` method that searches the augmented wavltree of mappings for a suitable gap given a `Layout`. The resulting spot will be randomized, but left-leaning.
Further method accepts an `entropy_bits` parameter that can be used to tune the sparseness of produced allocations. More consideration for this parameter can be found in the docs (docs/aslr.md)