refactor: new hardware memory management (#158)
* refactor: use xmas-elf instead of object

* refactor(loader): remove `alloc` dependency

This change removes dependency on `alloc` from the loader, to archive this we removed the lz4 compression feature, use `xmas-elf` instead of `object` and don't use `panic-abort`.

This chage makes the loader slimmer and helps with upcoming refactoring

* refactor: remove `panic-abort` crate

With the previous change `panic-abort` has now become obsolete

* refactor: remove `elf` feature from kmm and move into loader

In preparation for the virtmem overhaul, this change moves the elf mapping functionality from the `kmm` crate to the loader.

* cleanup

* wip

* feat: `Riscv64Sv39::from_active`

This change allows managing an existing page table through the pmm interface

* feat(sync): `Once::try_call_once`

This change adds another method to the `Once` primitive that allows fallible closures

* chore: debug printing for page tables

This change allows `Riscv64Sv39` to print its page table when displayed using the `Display` trait

* chore: bring back kernel parsing

* refactor(loader): use `pmm` for memory management

* refactor: cleanup kernel for refactor

* refactor: reduce memory size for QEMU

* remove kmm

* fmt & fixes

* cleanup

* feat(loader): allocate & map kernel heap

* wip

* fix(loader): correctly report memory regions ptr and len

* wip(wavltree): impl Send and Sync for tree

* fmt

* refactor: move `vmm` crate into kernel

* wip

* refactor(pmm): remove overengineered arch trait

This change removes the admittedly overengineered `Arch` trait that caused viral generics throughout all dependents with simple functions and consts

* wavltree fixes

* cleanup

* fmt

* Update once.rs

* cleanup

* refactor: report FDT memory region through `BootInfo.memory_regions` array

* cleanup

* fix: report FDT through memory regions array

* fix: correctly handle unmapping

* clippy

* fix: use virtual FDT address is test runner

* refactor: rename `AddressSpace:: phys_offset` to `AddressSpace:: physical_memory_offset `

* docs

* fmt