Feat(kernel/vm): allocate virtual memory spot (#165)
* feat(wavltree): implement lifecycle hooks

* feat(kernel/vm): implement efficient gap search state

This change uses the `wavltree` lifecycle hooks to maintain an augmented version of the WAVLTree that supports more efficient gap search. This is inspired by a similar inspiration in the Zircon kernel and works by tracking the largest and smallest address of each nodes subtree, as well as the biggest gap in each subtree.

* feat(kernel/vm): use prng for ASLR

* refactor: remove ambiguous and broken `cursor`/`cursor_mut` in favor of `root`/`root_mut` methods (#164)

* chore(wavltree): release v0.0.1 (#161)

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* chore(deps): update rust crate target-lexicon to 0.13.0 (#157)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* feat(kernel): parse FDT on startup

* fix: seed ASLR prng from provided seed

* feat(kernel/vm): allocate virtual memory spot

Implement the `find_spot` method that searches the augmented wavltree of mappings for a suitable gap given a `Layout`. The resulting spot will be randomized, but left-leaning.
Further method accepts an `entropy_bits` parameter that can be used to tune the sparseness of produced allocations. More consideration for this parameter can be found in the docs (docs/aslr.md)

* wip: test the pick_spot impl

* clippy & fmt

* refactor: move find_spot tests into tests

* fmt

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>