refactor(loader): use `BuddyAllocator` instead of `BumpAllocator`
This change makes the loader use the new `BuddyAllocator` instead of the `BumpAllocator`. The main reason is to benefit from the better alignment handling of the BuddyAllocator when allocating memory for the kernel stacks, TLS region, and heap so that super/hugepages can be used more often.

Previously we would fall back often on  mapping using regular 4k pages which is quite inefficient for e.g. the kernel heap.