feat: use buddy alloc (#176)
* refactor(pmm): `FrameAllocator` takes `Layout` as argument

This change allows users of `FrameAllocator` to pass a desired alignment in addition to the size of allocation. This alignment *must* be respected by allocator implementation.

* refactor(loader-api): Add `BootInfo::memory_regions` convenience method

This change adds a convenience `memories` method to `BootInfo` that allows accessing the slice of MemoryRegions without having to manually construct it every time.

* refactor(loader): use `BuddyAllocator` instead of `BumpAllocator`

This change makes the loader use the new `BuddyAllocator` instead of the `BumpAllocator`. The main reason is to benefit from the better alignment handling of the BuddyAllocator when allocating memory for the kernel stacks, TLS region, and heap so that super/hugepages can be used more often.

Previously we would fall back often on  mapping using regular 4k pages which is quite inefficient for e.g. the kernel heap.

* wip(kernel): use `BuddyAllocator` in kernel virtmem handling

This change makes the kernel virtmem handling code use the new buddy allocator instead of the `BitMap` one, to allow us to make better use of aligments in the future.

* refactor: move arch virtmem init routine into arch module

* refactor: remove BitMap and Bump frame alloocators