feat(thread-local): Better thread-local impl
This change upgrades the thread-local implementation used, the big reason is discovered UB in the way `Option` was used to detect uninitialized fields (the layout of `Option<T>` is actually *not* required to produce `Option::<T>::None` when "transmuting" from an zeroed byte slice).

This change also takes the opportunity and adds proper destructor  (ie `Drop`) support to thread local variables.