feat(kernel): better frame allocator (#208)
* refactor: rework startup sequence

* Update main.rs

* chore(loader): report TLS template alignment

This change ammends the `TlsTemplate` struct with a field reporting the chosen alignment for the segment so that the kernel may correctly align allocations. Currently the chosen alignment is much lower (8 byte) than the minimum allocation alignment 4096 bytes, but in the future this might change.

* feat(thread-local): Better thread-local impl

This change upgrades the thread-local implementation used, the big reason is discovered UB in the way `Option` was used to detect uninitialized fields (the layout of `Option<T>` is actually *not* required to produce `Option::<T>::None` when "transmuting" from an zeroed byte slice).

This change also takes the opportunity and adds proper destructor  (ie `Drop`) support to thread local variables.

* fix(kernel): reenable time module

* fixes

* clippy & fmt

* fix: disable tests

* feat(kernel): frame allocator

This change implements a proper frame allocator for the kernel, that *unlike* the previous frame allocation schemes primarily deals with `Frame` metadata structs that will be used throughout the higher-level virtual memory management system

* Update mod.rs

* fixes, clippy & fmt

* fmt & cargo deny

* refactor

* refactor: use `NonNull<T>` instead of `Pin<Unique<T>>`

* refactor

* fmt