feat(fiber): add support for forced fiber unwind (#417)
* feat: panic-unwind

This change extracts the kernel's panic.rs logic into a standalone crate (`panic-unwind`) so it can be shared with other library crates such as `fiber`.

* add better system startup docs

* feat(fiber): add support for forced fiber unwind

This change adds support for forcefully unwinding a fiber. This works by unwinding all the way through the fiber stack
and into the callers stack where it will automatically be caught (since we don't actually want to propagate this panic further). 
This helps to ensure that the fiber is correctly cleaned up and all `Drop` impls are correctly called along the way.