refactor: integrate async-exec into kernel (#458)
* integrate async-exec into kernel

* bring back shell

* fix: don't panic on zero stolen tasks

* fix: remove unecessary unparks from trap handler

* fix: remove park/unpark protocol

The protocol previously implemented by Parker & UnparkToken was designed to prevent spurious wakeups. However, spurious wakeups are not a problem for schedulers
since the queues are a simple way of determining whether the wakeup was spurious or not. And indeed it is desireable to check the queue in these cases, as the 
interrupt that caused the wakeup might have unblocked a task.