wip(kmem): implement frame allocator
This change adds the `FrameAllocator` which maintains a set of `Area`s which each manages a contiguous region of physical memory (may including holes).
`FrameAllocator` is Sync and holds the areas behind a spinlock. To reduce lock contention it keeps a CPU-local cache of up to 256 frames that will be 
allocated from first.

Each `Area` is a buddy-allocator that can allocate contiguous chunks with a power-of-two size and alignment up to a maximum size or alignment of 2^11.

The area selection algorithm, implemented in `frame_allocators/area_selection.rs` will choose the size of each area to minize the amount of wasted space
due to bookkeeping overhead and holes.