refactor: replace `checked_` methods
This change replaces the `checked_` methods previously exposed by the address types with regular `add`/`sub` methods that will panic in debug mode and _would_ wrap in release mode.
They do not wrap in release mode however because this change also enables overflow checks across the board. The intention here is to bring the behaviour of the address types more in line with the behaviour of other integer types used in the kernel. Knobs to control overflow checks should apply to both.