refactor(BootstrapAllocator): fix unaligned allocations & clean up implementation
The existing `BootstrapAllocator` implementation was less than ideal: It did not return correctly aligned blocks, its internal implementation with the cross-region offset
was hard to understand and debug and lastly it was potentially wasting physical memory by "jumping"the offset to the next region.

This change completely overhauls the implementation, now featuring a list of `Arenas` that each manage a contiguous physical memory region and hold their own bump pointers.
This is both much easier to understand and produces actually correct allocations (including cleaning up on partial faliures in `allocate`)